Fix menu Glitch On first Click

+++ app/scripts/controllers/menuCtrl.js
@@ -5,6 +5,7 @@ angular.module("mainSite")
 
   var docElemOrBody = document.documentElement || document.body;
   var scrollPostition = window.pageYOffset;
+  var justLoaded = 1;
   $scope.home = function() {
     window.scrollTo(0,0);
     menuService.removeOtherActivesOnClick("bg");
@@ -15,24 +16,48 @@ angular.module("mainSite")
     window.scrollTo(0, yCord);
   }
   $scope.skills = function() {
-    scrollToFun(document.getElementById("PUTBACKGROUNDHERE"));
-    menuService.removeOtherActivesOnClick("PUTBACKGROUNDHERE");
-    menuCollapseService.collapseMenu();
+    if (justLoaded == 1) {
+      scrollToFun(document.getElementById("PUTBACKGROUNDHERE"));
+      justLoaded = 0;
+    }
+    else {
+      scrollToFun(document.getElementById("PUTBACKGROUNDHERE"));
+      menuService.removeOtherActivesOnClick("PUTBACKGROUNDHERE");
+      menuCollapseService.collapseMenu();
+    }
   }
   $scope.bio = function() {
-    scrollToFun(document.getElementById("PUTBACKGROUNDHERE2"));
-    menuService.removeOtherActivesOnClick("PUTBACKGROUNDHERE2");
-    menuCollapseService.collapseMenu();
+    if (justLoaded == 1) {
+      scrollToFun(document.getElementById("PUTBACKGROUNDHERE"));
+      justLoaded = 0;
+    }
+    else {
+      scrollToFun(document.getElementById("PUTBACKGROUNDHERE2"));
+      menuService.removeOtherActivesOnClick("PUTBACKGROUNDHERE2");
+      menuCollapseService.collapseMenu();
+    }
   }
   $scope.port = function() {
-    scrollToFun(document.getElementById("PUTBACKGROUNDHERE3"));
-    menuService.removeOtherActivesOnClick("PUTBACKGROUNDHERE3");
-    menuCollapseService.collapseMenu();
+    if (justLoaded == 1) {
+      scrollToFun(document.getElementById("PUTBACKGROUNDHERE"));
+      justLoaded = 0;
+    }
+    else {
+      scrollToFun(document.getElementById("PUTBACKGROUNDHERE3"));
+      menuService.removeOtherActivesOnClick("PUTBACKGROUNDHERE3");
+      menuCollapseService.collapseMenu();
+    }
   }
   $scope.contact = function() {
-    scrollToFun(document.getElementById("PUTBACKGROUNDHERE4"));
-    menuService.removeOtherActivesOnClick("PUTBACKGROUNDHERE4");
-    menuCollapseService.collapseMenu();
+    if (justLoaded == 1) {
+      scrollToFun(document.getElementById("PUTBACKGROUNDHERE"));
+      justLoaded = 0;
+    }
+    else {
+      scrollToFun(document.getElementById("PUTBACKGROUNDHERE4"));
+      menuService.removeOtherActivesOnClick("PUTBACKGROUNDHERE4");
+      menuCollapseService.collapseMenu();
+    }
   }
 
   $timeout(function() {
